Moved subscriptions to graphql directory
