+++ server.js
@@ -5,7 +5,8 @@ import cors from "cors";
 import {createServer} from "http";
 import {SubscriptionServer} from "subscriptions-transport-ws";
 import {printSchema} from "graphql/utilities/schemaPrinter";
-import {subscriptionManager} from "./subscriptions/subscriptions";
+import {mongoConn} from './db/connect'
+import {subscriptionManager} from "./graphql/subscriptions/subscriptions";
 import schema from "./graphql/schema/schema";
 
 const GRAPHQL_PORT = 8080;
